Find the location of template data files on your system.

Find the location of template data files on your system.

Usage

autocv_resources(path = NULL, inst_dir = c("extdata", "templates"))

See also

Examples

# Obtain a complete list of top-level resources included with autocv:
autocv_resources()
#> [1] "css"       "latex"     "resources"

# Obtain a complete list of resources included in the css/ directory:
autocv_resources("css")
#> [1] "custom_resume.css" "styles_html.css"   "styles_pdf.css"   

# Obtain a complete path to the custom_resume.css resource:
autocv_resources("css/custom_resume.css")
#> [1] "/home/runner/.cache/R/renv/library/autocv-e7350c3b/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/autocv/extdata/css/custom_resume.css"

# Obtain a complete list of template files included with autocv:
autocv_resources(inst_dir = "templates")
#> [1] "template_cover_data.xlsx"  "template_cv.Rmd"          
#> [3] "template_job_metadata.yml" "template_resume.Rmd"      
#> [5] "template_resume_data.xlsx"