Create the file tree and data files for the present application.

Create the file tree and data files for the present application.

Usage

build_app_directory(app_df = construct_app_metadata(), open = TRUE)

See also

Examples

# Complete end-to-end example (build -> (edit) -> render -> check) ----------
in_tmp_env({
  message("1. Building project directory...")
  build_base_directory()
  
  data(example_job_metadata)
  app_df <- construct_app_metadata(app_info = example_job_metadata)
  
  message("")
  message("2. Building application...")
  build_app_directory(app_df = app_df, open = FALSE)
  
  message("")
  message("3. Rendering application...")
  render_app(cover = FALSE, email = FALSE)
  
  message("")
  message("4. Checking keywords...")
  report_df <- check_skills()
  print(report_df)
})
#> 1. Building project directory...
#> 
#> ── Setting project paths ───────────────────────────────────────────────────────
#> 
#> Your current path is: /tmp/RtmpqhwSrS/
#> Set your desired project path relative to your current path: 
#> 
#> You entered the path: /tmp/RtmpqhwSrS/
#> 
#> Setting your project root to: /tmp/RtmpqhwSrS/
#> Created file .here in /tmp/RtmpqhwSrS . Please start a new R session in the new project directory.
#> ✔ Writing file: .Rprofile
#> ✔ Writing file: mycv.Rproj
#> 
#> ℹ Writing to: ../../../../../../../tmp/RtmpqhwSrS
#> 
#> ── Creating directory tree ─────────────────────────────────────────────────────
#> ! Folder already exists: . (skipping) 
#> ✔ Creating folder: R
#> ✔ Creating folder: R/input
#> ✔ Creating folder: R/output
#> ✔ Creating folder: R/applications
#> 
#> ── Writing skeleton data files ─────────────────────────────────────────────────
#> ✔ Writing file: R/input/resume_data.xlsx
#> ✔ Writing file: R/input/cover_data.xlsx
#> ✔ Writing file: R/input/job_metadata.yml
#> 
#> ── Writing resume-building notebooks ───────────────────────────────────────────
#> ✔ Writing file: resume.Rmd
#> ✔ Writing file: cv.Rmd
#> 
#> 2. Building application...
#> 
#> ℹ Writing to: ../../../../../../../tmp/RtmpqhwSrS/R/applications/2024-07-data-science/2024-08-15-01-company-name-data-scientist-AB
#> 
#> ── Creating directory tree ─────────────────────────────────────────────────────
#> ✔ Creating folder: input
#> ✔ Creating folder: output
#> 
#> ── Copying base data files into directory ──────────────────────────────────────
#> ✔ Writing file: input/resume_data_AB.xlsx
#> ✔ Writing file: input/cover_data_AB.xlsx
#> 
#> ── Writing application metadata ────────────────────────────────────────────────
#> ✔ Writing file: metadata_AB.yml
#> ✔ Writing file: ../log.rds
#> 
#> ── Downloading job posting and building a skill report ─────────────────────────
#> ✔ Writing file: input/posting_AB.txt
#> 
#> ── Keyword check: Posting vs job terms list ────────────────────────────────────
#> ✔ Writing file: output/keyword_counts_AB.csv
#> # A tibble: 1 × 2
#>   term     matches
#>   <chr>      <dbl>
#> 1 Disabled       4
#> 
#> ── Keyword check: Posting vs data terms list ───────────────────────────────────
#> ✔ Writing file: output/skill_counts_posting_AB.csv
#> # A tibble: 55 × 2
#>    term                        matches
#>    <chr>                         <dbl>
#>  1 Data Science                     73
#>  2 Data Analysis                    15
#>  3 UX                                1
#>  4 Statistics                       29
#>  5 Cloud Computing                   4
#>  6 Insights                         10
#>  7 Computer Science                  7
#>  8 Survey Data                       1
#>  9 Algorithms                        4
#> 10 Unstructured Data                 2
#> 11 Domain Knowledge                  3
#> 12 Mathematics                       3
#> 13 Data-Driven                       3
#> 14 Statistical Knowledge             1
#> 15 Data Visualization                3
#> 16 Machine Learning                  8
#> 17 C                                 8
#> 18 Classification                    3
#> 19 Data Mining                       4
#> 20 Statistical Learning              3
#> 21 Data Collection                   1
#> 22 Big Data                         13
#> 23 Exploratory Data Analysis         2
#> 24 Data Management                   2
#> 25 Data Cleaning                     1
#> 26 Predictive Models                 3
#> 27 Models                            2
#> 28 Data-Driven Decisions             1
#> 29 Statistical Analysis              2
#> 30 Data Preprocessing                1
#> 31 Feature Engineering               1
#> 32 Model Selection                   1
#> 33 Implementation                    1
#> 34 Analytical Techniques             1
#> 35 Analytical                        1
#> 36 Large Datasets                    2
#> 37 Non-Technical Audiences           1
#> 38 Critical Thinking                 1
#> 39 Data-Driven Decision-Making       1
#> 40 Data Analytics                    1
#> 41 Cloud Services                    1
#> 42 Python                            1
#> 43 R                                 2
#> 44 Data Engineering                  1
#> 45 HTML                              4
#> 46 Statistical Modeling              1
#> 47 Causal Inference                  1
#> 48 Collaboration                     1
#> 49 Signal Processing                 1
#> 50 Spark                             1
#> 51 Spark SQL                         1
#> 52 Data Ethics                       1
#> 53 ETL                               1
#> 54 ELT                               1
#> 55 Developers                        1
#> 
#> ✔ Opening: .
#> 
#> 3. Rendering application...
#> 
#> ── Building resume.txt ─────────────────────────────────────────────────────────
#> ✔ Writing file: output/resume_yourname_AB.txt
#> 
#> ── Building resume.pdf ─────────────────────────────────────────────────────────
#> ✔ Writing file: output/resume_yourname_AB.pdf
#> ✔ All fields are up-to-date. Synchronizing metadata and log...
#> 
#> ! File already exists: metadata_AB.yml (overwriting) 
#> ! File already exists: ../log.rds (updating entry for 'AB') 
#> ! Package fontspec Error: 
#> (fontspec)                The font "Helvetica Neue" cannot be found; this
#> (fontspec)                may be but usually is not a fontspec bug. Either
#> (fontspec)                there is a typo in the font name/file, the font is
#> (fontspec)                not installed (correctly), or there is a bug in
#> (fontspec)                the underlying font loading engine
#> (fontspec)                (XeTeX/luaotfload).
#> Error: LaTeX failed to compile /tmp/RtmpqhwSrS//./R/applications/2024-07-data-science/2024-08-15-01-company-name-data-scientist-AB/output/resume_yourname_AB.tex. See https://yihui.org/tinytex/r/#debugging for debugging tips. See resume_yourname_AB.log for more info.