Create the file tree and necessary data files to begin building a resume.

Create the file tree and necessary data files to begin building a resume.

Usage

build_base_directory()

See also

Examples

in_tmp_env({
  build_base_directory()
})
#> 
#> ── Setting project paths ───────────────────────────────────────────────────────
#> 
#> Your current path is: /tmp/RtmpxfDVOm/
#> Set your desired project path relative to your current path: 
#> 
#> You entered the path: /tmp/RtmpxfDVOm/
#> 
#> Setting your project root to: /tmp/RtmpxfDVOm/
#> File .here already exists in /tmp/RtmpxfDVOm
#> ! File already exists: .Rprofile (skipping) 
#> ! File already exists: mycv.Rproj (skipping) 
#> 
#> ℹ Writing to: ../../../../../../../tmp/RtmpxfDVOm
#> 
#> ── Creating directory tree ─────────────────────────────────────────────────────
#> ! Folder already exists: . (skipping) 
#> ! Folder already exists: R (skipping) 
#> ! Folder already exists: R/input (skipping) 
#> ! Folder already exists: R/output (skipping) 
#> ! Folder already exists: R/applications (skipping) 
#> 
#> ── Writing skeleton data files ─────────────────────────────────────────────────
#> ! File already exists: R/input/resume_data.xlsx (skipping) 
#> ! File already exists: R/input/cover_data.xlsx (skipping) 
#> ! File already exists: R/input/job_metadata.yml (skipping) 
#> 
#> ── Writing resume-building notebooks ───────────────────────────────────────────
#> ! File already exists: resume.Rmd (skipping) 
#> ! File already exists: cv.Rmd (skipping)