Sample contact data

A synthetic dataset containing personal contact information and associated links, including a home address, email, and personal websites. For more details, refer to the corresponding template spreadsheet in the repository.

Usage

example_contact_data

Format

A data frame with 9 rows and 5 columns:

loc

Name of the contact info field (character).

order

Ordinal position in which to display the item in your resume (positive integer).

icon

Name of a Font Awesome icon to display next to the item; optional (character).

address_text

Text to display for the item, in place of the address, if provided (see below) (character).

address

Link or path to the website, web address, or file to be redirected to when the corresponding address_text is clicked; optional (character).

Source

Generated for demonstration purposes.

Examples

data(example_contact_data)
head(example_contact_data)
#> # A tibble: 6 × 5
#>   loc      order icon           address_text        address                   
#>   <chr>    <dbl> <chr>          <chr>               <chr>                     
#> 1 name         0 NA             My Name             NA                        
#> 2 title       NA NA             PhD                 NA                        
#> 3 pic         NA NA             My Name             /absolute/path/to/pic.png 
#> 4 location     1 map-marked-alt City, ST            NA                        
#> 5 email        2 envelope       myemail@outlook.com mailto:myemail@outlook.com
#> 6 phone        3 phone          (999) 999-9999      tel:+19999999999