Sample job metadata

A synthetic dataset containing information about a hypothetical job. For more details, refer to the corresponding template file in the repository.

Usage

example_job_metadata

Format

A named list with 11 fields:

name

Your name (character).

period

Name of the current job application period, chosen by the user when creating the first application for the period; must be a valid directory name (character).

id

Unique identifier (relative to all other identifiers for the current period) to associate with the current application (character).

base_id

Identifier corresponding to an existing job application upon which to base the current application. If provided, the template data files for the current application directory will be copied from the base_id directory rather than the autocv template directory; must be an existing id for the current period; optional (character).

company

Name of the company to which you are applying (character).

position

Name of the position for which you are applying (character).

portal_url

URL of the job application portal; optional (character).

posting_url

URL of the job posting page (character).

linkedin_url

URL of the LinkedIn posting for the job application; optional (character).

recruiter_email

Email of the recruiter for the job; optional (character).

notes

Any notes to include in the log for the current job application; optional (character).

Source

Generated for demonstration purposes.

Examples

data(example_job_metadata)
head(example_job_metadata)
#> $name
#> [1] "Your Name"
#> 
#> $period
#> [1] "2024-07-data-science"
#> 
#> $id
#> [1] "AB"
#> 
#> $base_id
#> [1] "/"
#> 
#> $company
#> [1] "Company Name"
#> 
#> $position
#> [1] "Data Scientist"
#>