Sample position data

A synthetic dataset containing information about ...

Usage

example_position_data

Format

A data frame with x rows and y columns:

field

description

skill_1, skill_2, ..., skill_5

description (all character)

skill_ix_iy

, where ix and iy can each be an integer from 1 to 5 (all character)

Source

Generated for demonstration purposes.

Examples

data(example_position_data)
head(example_position_data)
#> # A tibble: 6 × 35
#>   section        title     in_base include institution loc   start              
#>   <chr>          <chr>     <chr>   <chr>   <chr>       <chr> <dttm>             
#> 1 certifications Certific… x       x       Institutio… Loca… 1900-01-01 00:00:00
#> 2 certifications Certific… x       x       Institutio… Loca… 1900-01-01 00:00:00
#> 3 education      PhD, Deg… x       x       University… City… 2021-09-01 00:00:00
#> 4 education      BSc, Deg… x       x       University… City… 2014-09-01 00:00:00
#> 5 education      MSc, Deg… x       x       University… City… 2018-09-01 00:00:00
#> 6 projects       pkgname   x       x       NA          NA    2024-01-01 00:00:00
#> # ℹ 28 more variables: end <dttm>, link <chr>, link_text <chr>,
#> #   short_summary <chr>, description_1 <chr>, description_2 <chr>,
#> #   description_3 <chr>, description_4 <chr>, description_5 <lgl>,
#> #   skill_1_1 <chr>, skill_1_2 <chr>, skill_1_3 <chr>, skill_1_4 <chr>,
#> #   skill_1_5 <lgl>, skill_2_1 <lgl>, skill_2_2 <lgl>, skill_3_1 <lgl>,
#> #   skill_3_2 <lgl>, skill_4_1 <lgl>, skill_4_2 <lgl>, skill_5_1 <lgl>,
#> #   skill_5_2 <lgl>, tool_1 <lgl>, tool_2 <lgl>, tool_3 <lgl>, …