Sample job posting

A hypothetical job posting.

Usage

example_posting

Format

A character vector of length 1

Source

Generated for demonstration purposes.

Examples

data(example_posting)
head(example_posting)
#> [1] "Data Scientist, Management: People Analytics"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [2] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> [3] "At autocv, our mission is to improve peoples lives with the worlds best transportation. To do this, we start with our own community by creating an open, inclusive, and diverse organization. As a Data Scientist on the Driving XP team, you will collaborate with our world class team of engineers, product managers, and designers to design the end to end driving experience for hundreds of thousands of drivers who interact deeply with our app every day. This includes building tools to set personal driving preferences, guidance on when and where to drive, and designing the best UI and personalized XP to get the right rides to the right drivers quickly. Data Science is at the heart of autocvs products and decision-making. You will leverage data and rigorous, analytical thinking to shape our Driving XP products and make business decisions that put drivers first. This will involve identifying and scoping opportunities, shaping priorities, recommending technical solutions, designing experiments, and measuring the impact of new features. Responsibilities: Leverage data and analytic frameworks to identify opportunities for growth and efficiency Partner with product managers, engineers, marketers, designers, and operators to translate data insights into decisions and actionDesign and analyze online experiments; communicate results and act on launch decisionsDevelop analytical frameworks to monitor business and product performance Establish metrics that measure the health of our products, as well as rider and driver experience Experience: Degree in a quantitative field such as statistics, economics, applied math, operations research or engineering (advanced degrees preferred), or relevant work experience (3+ years of industry experience in a data science or analytical role) Proficiency in SQL (able to write structured and efficient queries on large data sets) Experience in programming, especially with data science and visualization libraries in Python or R Experience in online experimentation and statistical analysis Strong oral and written communication skills, and ability to collaborate with and influence cross-functional partners Benefits: Extended health and dental coverage options, along with life insurance and disability benefits Mental health benefits Family building benefits Access to a Health Care Savings Account In addition to provincial observed holidays, team members get 15 days paid time off, with an additional day for each year of service 4 Floating Holidays each calendar year prorated based off of date of hire 10 paid sick days per year regardless of province 18 weeks of paid parental leave. Biological, adoptive, and foster parents are all eligible autocv proudly pursues and hires a diverse workforce. This role will be in-office on a hybrid schedule. Team Members will be expected to work in the office 3 days per week. Additionally, hybrid roles have the flexibility to work from anywhere for up to 4 weeks per year."