Sample text data

A synthetic dataset containing the textual building blocks of job application components such as a resume bio or a cover letter.

Usage

example_text_data

Format

A data frame with 32 rows and 7 columns:

loc

Name of the text block (character).

include

Flag indicating whether the item should be included in your tailored resume, either "x" or NA (character).

text

The actual text to display for the item (character).

order

Ordinal position in which to display the item in your application documents (positive integer).

word_limit

Limit to impose on the number of words for the text block (positive integer).

word_count

Computed number of words for the text block; auto-populated (positive integer).

Source

Generated for demonstration purposes.

Examples

data(example_text_data)
head(example_text_data)
#> # A tibble: 6 × 7
#>   loc            include text       order word_limit word_count `notes/outtakes`
#>   <chr>          <chr>   <chr>      <dbl>      <dbl>      <dbl> <chr>           
#> 1 title          x       My Title      NA         15          2 NA              
#> 2 bio_experience x       This is m…     1         15          5 NA              
#> 3 bio_expertise  NA      This is m…     2         15          5 NA              
#> 4 bio_value      x       This is m…     3         15          5 NA              
#> 5 bio_objective  NA      This is m…     4         15          5 NA              
#> 6 bio_experience NA      /              1         15          1 NA