Get functions

Functions to get resources and example data built into autoCV.

get_template(save_dir = here::here())

get_data(dir_only = TRUE, dir_manual = NULL)

get_css()

get_logo(
  width = "40px",
  id = "header_img",
  text_before = NULL,
  text_after = NULL,
  link = "https://github.com/bschilder/autoCV",
  img = system.file("hex", "hex.png", package = "autoCV"),
  style = "border-radius:5%;"
)

Arguments

save_dir

Directory to save template file to.

dir_only

Only return the directory.

dir_manual

Specify the directory to find the files in manually.

width

Image width.

id

HTML element id.

text_before

HTML element text before image.

text_after

HTML element text after image.

link

URL to use.

img

Path or URL to image.

style

HTML element style.

Value

Resources.

Functions

get_template(): get_

get_data(): get_

get_css(): get_

get_logo(): get_