Construct an absolute path to a project directory for saving and loading.

This function constructs an absolute path to a specified project directory. Directory paths, relative to the project root, are defined in the user's .Rprofile environment variables.

Usage

get_path_to(
  dir = c("src", "input", "output", "applications", "notebooks", "templates", "css",
    "latex", "resources")
)

Arguments

dir

(string) the directory name.

Examples

get_path_to("src")
#> [1] "/home/runner/work/autocv/autocv/./R"
get_path_to("templates")
#> [1] "/home/runner/.cache/R/renv/library/autocv-e7350c3b/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/autocv/templates"