Open, edit, and remove application documents.

open_app opens documents (resume, by default) for a given application.

edit_app opens spreadsheet data files for a given application for editing.

edit_base opens base spreadsheet data files for editing.

delete_app deletes the application folder and log entry for a given id.

Usage

open_app(doc = "resume", id = "latest", app_period = "latest")

edit_app(id = "latest", app_period = "latest")

edit_base()

delete_app(app_id, log = load_log())

See also