Prepare raw strings and entries in a keyword list for analysis.

Strip markdown characters, trailing whitespace, blank lines, and all but the unique lines from the supplied term list.

Usage

prep_term_list(terms)

Arguments

terms

A character vector containing keywords.

Value

A character vector containing only the unique, processed keywords.

See also

Examples

terms <- c("### ML", "", "#### Supervised Learning", "SVMs", "ML")
print(terms)
#> [1] "### ML"                   ""                        
#> [3] "#### Supervised Learning" "SVMs"                    
#> [5] "ML"                      

processed_terms <- prep_term_list(terms)
print(processed_terms)
#> [1] "ML"                  "Supervised Learning" "SVMs"