Sample position data

A synthetic dataset containing information about previous and current professional roles and experiences, including work, education, certifications, projects, and publications. For more details, refer to the corresponding template spreadsheet in the repository.

Usage

example_position_data

Format

A data frame with 11 rows and 35 columns:

section

Name of the section under which the entry should appear in the resume (character).

title

Name or title of the role, experience, or project (character).

in_base

Flag indicating whether the entry should be included in your base resume, either "x" or NA (character).

include

Flag indicating whether the entry should be included in your tailored resume, either "x" or NA (character).

institution

Name of the institution associated with the role or experience (character).

loc

Name of the location associated with the role or experience (character).

start

Start date (YYYY-MM-DD) of the role or experience; set to 1900-01-01 (the origin) to indicate that the provided end date (see below) is a target (will be rendered as "Expected - end" (date).

end

End date (YYYY-MM-DD) of the role or experience; set to NA to indicate that the role is ongoing (will be rendered as "start - Present") (date).

link

Web address to display next to the position name; optional (character).

link_text

The (clickable) text to display in place of the web address, if link supplied (character).

short_summary

Concise summary of the role or experience and the associated responsibilities, intended for use in a professional profile such as LinkedIn (character).

description_i

Bullet point to include under the entry, where i can be an integer from 1 to 5 representing the bullet's ordering (all character).

skill_ix_iy

Name of a (comma-separated) skill to include in parentheses next to bullet point ix, in the order given by iy, where ix and iy can each be an integer from 1 to 5 (all character).

tool_1, tool_2, tool_3

Name of a tool or technology required for the role or experience, intended for use in a professional profile such as LinkedIn (all character).

competency_1, competency_2, competency_3

Name of a skill (other than a tool or technology) required for the role or experience, intended for use in a professional profile such as LinkedIn (all character).

Source

Generated for demonstration purposes.

Examples

data(example_position_data)
head(example_position_data)
#> # A tibble: 6 × 35
#>   section        title     in_base include institution loc   start              
#>   <chr>          <chr>     <chr>   <chr>   <chr>       <chr> <dttm>             
#> 1 certifications Certific… x       x       Institutio… Loca… 1900-01-01 00:00:00
#> 2 certifications Certific… x       x       Institutio… Loca… 1900-01-01 00:00:00
#> 3 education      PhD, Deg… x       x       University… City… 2021-09-01 00:00:00
#> 4 education      BSc, Deg… x       x       University… City… 2014-09-01 00:00:00
#> 5 education      MSc, Deg… x       x       University… City… 2018-09-01 00:00:00
#> 6 projects       pkgname   x       x       NA          NA    2024-01-01 00:00:00
#> # ℹ 28 more variables: end <dttm>, link <chr>, link_text <chr>,
#> #   short_summary <chr>, description_1 <chr>, description_2 <chr>,
#> #   description_3 <chr>, description_4 <chr>, description_5 <lgl>,
#> #   skill_1_1 <chr>, skill_1_2 <chr>, skill_1_3 <chr>, skill_1_4 <chr>,
#> #   skill_1_5 <lgl>, skill_2_1 <lgl>, skill_2_2 <lgl>, skill_3_1 <lgl>,
#> #   skill_3_2 <lgl>, skill_4_1 <lgl>, skill_4_2 <lgl>, skill_5_1 <lgl>,
#> #   skill_5_2 <lgl>, tool_1 <lgl>, tool_2 <lgl>, tool_3 <lgl>, …