Get functions

Functions to get resources and example data built into autoCV.

get_template(
  name = "Leonardo da Vinci",
  tagline = "Simplicity is the ultimate sophistication.",
  extra = list("### Studio of Andrea del Verrocchio",
    "### Painter, draughtsman, engineer, scientist", "### MD, PhD"),
  logo = get_logo(img = system.file("img", "davinci.png", package = "autoCV"), width =
    "60px"),
  wd = get_data(),
  template = "CV.Rmd",
  save_dir = "./",
  force_new = FALSE
)

get_data(
  dir_only = TRUE,
  dir_manual = NULL,
  save_dir = NULL,
  subdir = c("cv_data", "img", "css"),
  verbose = TRUE
)

get_data_list(wd, full.names = FALSE)

get_css()

get_logo(
  width = "40px",
  id = "header_img",
  text_before = NULL,
  text_after = NULL,
  link = "https://github.com/bschilder/autoCV",
  img = system.file("hex", "hex.png", package = "autoCV"),
  style = "border-radius:5%;"
)

Arguments

name

Image name.

tagline

Tagline to use.

extra

Extra HTML/markdown elements to append.

logo

Logo to use.

wd

Directory where "cv_data" and "img" subdirectories are stored.

template

Template to use.

save_dir

If not NULL, copy example data ("cv_data/"), images ("img/"), and CSS ("css/") to this directory.

force_new

Overwrite existing template file.

dir_only

Only return the directory.

dir_manual

Specify the directory to find the files in manually.

subdir

Subdirectories to return (when dir_only=FALSE).

verbose

Print messages.

full.names

a logical value. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.

width

Image width.

id

HTML element id.

text_before

HTML element text before image.

text_after

HTML element text after image.

link

URL to use.

img

Path or URL to image.

style

HTML element style.

Value

Resources.

Functions

get_template(): get_

get_data(): get_

get_data_list(): get_

get_css(): get_

get_logo(): get_

Examples

wd <-get_data(save_dir = file.path(tempdir(),"example_data"))
#> Copying example data --> /tmp/Rtmpg7Fl4k/example_data